+++ tests/dsp/platform.cpp
@@ -1397,12 +1397,12 @@ TEST(PlatformEventMessageSupported, testBadDecodeRespond)
     EXPECT_EQ(rc, PLDM_ERROR_INVALID_LENGTH);
 }
 
-TEST(PollForPlatformEventMessage, testGoodEncodeRequest)
+TEST(PollForPlatformEventMessage, testGoodEncodeRequestFirstPart)
 {
     uint8_t formatVersion = 0x01;
-    uint8_t transferOperationFlag = 0x1;
-    uint32_t dataTransferHandle = 0xffffffff;
-    uint16_t eventIdToAcknowledge = 0x0;
+    uint8_t transferOperationFlag = PLDM_GET_FIRSTPART;
+    uint32_t dataTransferHandle = 0xaabbccdd;
+    uint16_t eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_NULL;
 
     std::array<uint8_t,
                hdrSize + PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES>
@@ -1414,14 +1414,14 @@ TEST(PollForPlatformEventMessage, testGoodEncodeRequest)
         0, formatVersion, transferOperationFlag, dataTransferHandle,
         eventIdToAcknowledge, request,
         PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
-    EXPECT_EQ(rc, PLDM_SUCCESS);
+    EXPECT_EQ(rc, 0);
 
     struct pldm_msgbuf _buf;
     struct pldm_msgbuf* buf = &_buf;
-    rc = pldm_msgbuf_init_cc(
+    rc = pldm_msgbuf_init_errno(
         buf, PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES, request->payload,
         PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
-    EXPECT_EQ(rc, PLDM_SUCCESS);
+    EXPECT_EQ(rc, 0);
 
     uint8_t retFormatVersion;
     uint8_t retTransferOperationFlag;
@@ -1437,16 +1437,101 @@ TEST(PollForPlatformEventMessage, testGoodEncodeRequest)
     EXPECT_EQ(retTransferOperationFlag, transferOperationFlag);
     EXPECT_EQ(retDataTransferHandle, dataTransferHandle);
     EXPECT_EQ(retEventIdToAcknowledge, eventIdToAcknowledge);
-    EXPECT_EQ(pldm_msgbuf_destroy(buf), PLDM_SUCCESS);
+    EXPECT_EQ(pldm_msgbuf_destroy_consumed(buf), 0);
+}
+
+TEST(PollForPlatformEventMessage, testGoodEncodeRequestNextPart)
+{
+    uint8_t formatVersion = 0x01;
+    uint8_t transferOperationFlag = PLDM_GET_NEXTPART;
+    uint32_t dataTransferHandle = 0xaabbccdd;
+    uint16_t eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_FRAGMENT;
+
+    std::array<uint8_t,
+               hdrSize + PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES>
+        requestMsg{};
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    auto request = reinterpret_cast<pldm_msg*>(requestMsg.data());
+
+    auto rc = encode_poll_for_platform_event_message_req(
+        0, formatVersion, transferOperationFlag, dataTransferHandle,
+        eventIdToAcknowledge, request,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, 0);
+
+    struct pldm_msgbuf _buf;
+    struct pldm_msgbuf* buf = &_buf;
+    rc = pldm_msgbuf_init_errno(
+        buf, PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES, request->payload,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, 0);
+
+    uint8_t retFormatVersion;
+    uint8_t retTransferOperationFlag;
+    uint32_t retDataTransferHandle;
+    uint16_t retEventIdToAcknowledge;
+
+    pldm_msgbuf_extract_uint8(buf, &retFormatVersion);
+    pldm_msgbuf_extract_uint8(buf, &retTransferOperationFlag);
+    pldm_msgbuf_extract_uint32(buf, &retDataTransferHandle);
+    pldm_msgbuf_extract_uint16(buf, &retEventIdToAcknowledge);
+
+    EXPECT_EQ(retFormatVersion, formatVersion);
+    EXPECT_EQ(retTransferOperationFlag, transferOperationFlag);
+    EXPECT_EQ(retDataTransferHandle, dataTransferHandle);
+    EXPECT_EQ(retEventIdToAcknowledge, eventIdToAcknowledge);
+    EXPECT_EQ(pldm_msgbuf_destroy_consumed(buf), 0);
+}
+
+TEST(PollForPlatformEventMessage, testGoodEncodeRequestAckOnly)
+{
+    uint8_t formatVersion = 0x01;
+    uint8_t transferOperationFlag = PLDM_ACKNOWLEDGEMENT_ONLY;
+    uint32_t dataTransferHandle = 0xaabbccdd;
+    uint16_t eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_FRAGMENT;
+
+    std::array<uint8_t,
+               hdrSize + PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES>
+        requestMsg{};
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    auto request = reinterpret_cast<pldm_msg*>(requestMsg.data());
+
+    auto rc = encode_poll_for_platform_event_message_req(
+        0, formatVersion, transferOperationFlag, dataTransferHandle,
+        eventIdToAcknowledge, request,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, 0);
+
+    struct pldm_msgbuf _buf;
+    struct pldm_msgbuf* buf = &_buf;
+    rc = pldm_msgbuf_init_errno(
+        buf, PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES, request->payload,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, 0);
+
+    uint8_t retFormatVersion;
+    uint8_t retTransferOperationFlag;
+    uint32_t retDataTransferHandle;
+    uint16_t retEventIdToAcknowledge;
+
+    pldm_msgbuf_extract_uint8(buf, &retFormatVersion);
+    pldm_msgbuf_extract_uint8(buf, &retTransferOperationFlag);
+    pldm_msgbuf_extract_uint32(buf, &retDataTransferHandle);
+    pldm_msgbuf_extract_uint16(buf, &retEventIdToAcknowledge);
+
+    EXPECT_EQ(retFormatVersion, formatVersion);
+    EXPECT_EQ(retTransferOperationFlag, transferOperationFlag);
+    EXPECT_EQ(retDataTransferHandle, dataTransferHandle);
+    EXPECT_EQ(retEventIdToAcknowledge, eventIdToAcknowledge);
+    EXPECT_EQ(pldm_msgbuf_destroy_consumed(buf), 0);
 }
 
 TEST(PollForPlatformEventMessage, testBadEncodeRequest)
 {
     uint8_t formatVersion = 0x01;
-    uint8_t transferOperationFlag = 0x1;
-    uint32_t dataTransferHandle = 0xffffffff;
-    uint16_t eventIdToAcknowledge = 0x0;
-
+    uint8_t transferOperationFlag = PLDM_GET_FIRSTPART;
+    uint32_t dataTransferHandle = 0xaabbccdd;
+    uint16_t eventIdToAcknowledge = 0x1234;
     std::array<uint8_t,
                hdrSize + PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES>
         requestMsg{};
@@ -1457,12 +1542,55 @@ TEST(PollForPlatformEventMessage, testBadEncodeRequest)
         0, formatVersion, transferOperationFlag, dataTransferHandle,
         eventIdToAcknowledge, nullptr,
         PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, -EINVAL);
 
-    EXPECT_EQ(rc, PLDM_ERROR_INVALID_DATA);
-
-    encode_poll_for_platform_event_message_req(
+    transferOperationFlag = PLDM_GET_FIRSTPART;
+    eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_FRAGMENT;
+    rc = encode_poll_for_platform_event_message_req(
         0, formatVersion, transferOperationFlag, dataTransferHandle,
-        eventIdToAcknowledge, request, hdrSize);
+        eventIdToAcknowledge, request,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, -EPROTO);
+
+    transferOperationFlag = PLDM_GET_NEXTPART;
+    eventIdToAcknowledge = 0x1234;
+    rc = encode_poll_for_platform_event_message_req(
+        0, formatVersion, transferOperationFlag, dataTransferHandle,
+        eventIdToAcknowledge, request,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, -EPROTO);
+
+    transferOperationFlag = PLDM_GET_NEXTPART;
+    eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_NULL;
+    rc = encode_poll_for_platform_event_message_req(
+        0, formatVersion, transferOperationFlag, dataTransferHandle,
+        eventIdToAcknowledge, request,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, -EPROTO);
+
+    transferOperationFlag = PLDM_ACKNOWLEDGEMENT_ONLY;
+    eventIdToAcknowledge = 0x1234;
+    rc = encode_poll_for_platform_event_message_req(
+        0, formatVersion, transferOperationFlag, dataTransferHandle,
+        eventIdToAcknowledge, request,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, -EPROTO);
+
+    transferOperationFlag = PLDM_ACKNOWLEDGEMENT_ONLY;
+    eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_NULL;
+    rc = encode_poll_for_platform_event_message_req(
+        0, formatVersion, transferOperationFlag, dataTransferHandle,
+        eventIdToAcknowledge, request,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, -EPROTO);
+
+    transferOperationFlag = PLDM_ACKNOWLEDGEMENT_ONLY + 1;
+    eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_FRAGMENT;
+    rc = encode_poll_for_platform_event_message_req(
+        0, formatVersion, transferOperationFlag, dataTransferHandle,
+        eventIdToAcknowledge, request,
+        PLDM_POLL_FOR_PLATFORM_EVENT_MESSAGE_REQ_BYTES);
+    EXPECT_EQ(rc, -EPROTO);
 }
 
 TEST(PollForPlatformEventMessage, testGoodDecodeRespond)
@@ -1659,7 +1787,7 @@ TEST(PollForPlatformEventMessage, testGoodDecodeRequestFirstPart)
     uint8_t formatVersion = 0x1;
     uint8_t transferOperationFlag = PLDM_GET_FIRSTPART;
     uint32_t dataTransferHandle = 0x11223344;
-    uint16_t eventIdToAcknowledge = 0x0;
+    uint16_t eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_NULL;
     std::vector<uint8_t> requestMsg{0x1,  0x0,  0x0,  0x1,  PLDM_GET_FIRSTPART,
                                     0x44, 0x33, 0x22, 0x11, 0x00,
                                     0x00};
@@ -1676,7 +1804,7 @@ TEST(PollForPlatformEventMessage, testGoodDecodeRequestFirstPart)
         &retTransferOperationFlag, &retDataTransferHandle,
         &retEventIdToAcknowledge);
 
-    EXPECT_EQ(rc, PLDM_SUCCESS);
+    EXPECT_EQ(rc, 0);
     EXPECT_EQ(retFormatVersion, formatVersion);
     EXPECT_EQ(retTransferOperationFlag, transferOperationFlag);
     EXPECT_EQ(retDataTransferHandle, dataTransferHandle);
@@ -1688,7 +1816,7 @@ TEST(PollForPlatformEventMessage, testGoodDecodeRequestNextPart)
     uint8_t formatVersion = 0x1;
     uint8_t transferOperationFlag = PLDM_GET_NEXTPART;
     uint32_t dataTransferHandle = 0x11223344;
-    uint16_t eventIdToAcknowledge = 0xffff;
+    uint16_t eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_FRAGMENT;
     std::vector<uint8_t> requestMsg{0x1,  0x0,  0x0,  0x1,  PLDM_GET_NEXTPART,
                                     0x44, 0x33, 0x22, 0x11, 0xff,
                                     0xff};
@@ -1705,7 +1833,7 @@ TEST(PollForPlatformEventMessage, testGoodDecodeRequestNextPart)
         &retTransferOperationFlag, &retDataTransferHandle,
         &retEventIdToAcknowledge);
 
-    EXPECT_EQ(rc, PLDM_SUCCESS);
+    EXPECT_EQ(rc, 0);
     EXPECT_EQ(retFormatVersion, formatVersion);
     EXPECT_EQ(retTransferOperationFlag, transferOperationFlag);
     EXPECT_EQ(retDataTransferHandle, dataTransferHandle);
@@ -1717,7 +1845,7 @@ TEST(PollForPlatformEventMessage, testGoodDecodeRequestAck)
     uint8_t formatVersion = 0x1;
     uint8_t transferOperationFlag = PLDM_ACKNOWLEDGEMENT_ONLY;
     uint32_t dataTransferHandle = 0x11223344;
-    uint16_t eventIdToAcknowledge = 0xffff;
+    uint16_t eventIdToAcknowledge = PLDM_PLATFORM_EVENT_ID_FRAGMENT;
     std::vector<uint8_t> requestMsg{
         0x1,  0x0,  0x0,  0x1, PLDM_ACKNOWLEDGEMENT_ONLY, 0x44, 0x33,
         0x22, 0x11, 0xff, 0xff};
@@ -1734,7 +1862,7 @@ TEST(PollForPlatformEventMessage, testGoodDecodeRequestAck)
         &retTransferOperationFlag, &retDataTransferHandle,
         &retEventIdToAcknowledge);
 
-    EXPECT_EQ(rc, PLDM_SUCCESS);
+    EXPECT_EQ(rc, 0);
     EXPECT_EQ(retFormatVersion, formatVersion);
     EXPECT_EQ(retTransferOperationFlag, transferOperationFlag);
     EXPECT_EQ(retDataTransferHandle, dataTransferHandle);
@@ -1743,95 +1871,93 @@ TEST(PollForPlatformEventMessage, testGoodDecodeRequestAck)
 
 TEST(PollForPlatformEventMessage, testBadDecodeRequest)
 {
+    /*
+     * transfer_operation_flag is PLDM_GET_FIRSTPART and
+     * event_id_to_acknowledge is not PLDM_PLATFORM_EVENT_ID_NULL
+     */
     std::vector<uint8_t> requestMsg{0x1,  0x0,  0x0,  0x1,  PLDM_GET_FIRSTPART,
                                     0x44, 0x33, 0x22, 0x11, 0x66,
                                     0x55};
     // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
     auto request = reinterpret_cast<pldm_msg*>(requestMsg.data());
-
     uint8_t retFormatVersion;
     uint8_t retTransferOperationFlag;
     uint32_t retDataTransferHandle;
     uint16_t retEventIdToAcknowledge;
 
-    auto rc = decode_poll_for_platform_event_message_req(
-        NULL, requestMsg.size() - hdrSize, &retFormatVersion,
-        &retTransferOperationFlag, &retDataTransferHandle,
-        &retEventIdToAcknowledge);
-    EXPECT_EQ(rc, PLDM_ERROR_INVALID_DATA);
-
     /*
      * transfer_operation_flag is not PLDM_GET_FIRSTPART or PLDM_GET_NEXTPART or
      * PLDM_ACKNOWLEDGEMENT_ONLY
      */
-
     requestMsg[4] = PLDM_ACKNOWLEDGEMENT_ONLY + 1;
 
-    rc = decode_poll_for_platform_event_message_req(
+    auto rc = decode_poll_for_platform_event_message_req(
         request, requestMsg.size() - hdrSize, &retFormatVersion,
         &retTransferOperationFlag, &retDataTransferHandle,
         &retEventIdToAcknowledge);
 
-    EXPECT_EQ(rc, PLDM_ERROR_INVALID_DATA);
-    /*
-     * transfer_operation_flag is PLDM_GET_FIRSTPART and
-     * event_id_to_acknowledge not 0x0000
-     */
-    requestMsg[4] = PLDM_GET_FIRSTPART;
-    requestMsg[9] = 0x0;
-    requestMsg[10] = 0x1;
-
-    rc = decode_poll_for_platform_event_message_req(
-        request, requestMsg.size() - hdrSize, &retFormatVersion,
-        &retTransferOperationFlag, &retDataTransferHandle,
-        &retEventIdToAcknowledge);
-
-    EXPECT_EQ(rc, PLDM_ERROR_INVALID_DATA);
-
-    /*
-     * transfer_operation_flag is not PLDM_GET_FIRSTPART and
-     * event_id_to_acknowledge is 0x0000
-     */
-    requestMsg[4] = PLDM_GET_NEXTPART;
-    requestMsg[9] = 0x0;
-    requestMsg[10] = 0x0;
-
-    rc = decode_poll_for_platform_event_message_req(
-        request, requestMsg.size() - hdrSize, &retFormatVersion,
-        &retTransferOperationFlag, &retDataTransferHandle,
-        &retEventIdToAcknowledge);
-
-    EXPECT_EQ(rc, PLDM_ERROR_INVALID_DATA);
+    EXPECT_EQ(rc, -EPROTO);
 
     /*
      * transfer_operation_flag is PLDM_GET_NEXTPART and
-     * event_id_to_acknowledge not 0xffff
+     * event_id_to_acknowledge is not PLDM_PLATFORM_EVENT_ID_FRAGMENT
      */
     requestMsg[4] = PLDM_GET_NEXTPART;
-    requestMsg[9] = 0x0;
-    requestMsg[10] = 0x1;
+    requestMsg[9] = 0x11;
+    requestMsg[10] = 0x22;
 
     rc = decode_poll_for_platform_event_message_req(
         request, requestMsg.size() - hdrSize, &retFormatVersion,
         &retTransferOperationFlag, &retDataTransferHandle,
         &retEventIdToAcknowledge);
 
-    EXPECT_EQ(rc, PLDM_ERROR_INVALID_DATA);
+    EXPECT_EQ(rc, -EPROTO);
 
     /*
-     * transfer_operation_flag is not PLDM_GET_NEXTPART and
-     * event_id_to_acknowledge is 0xffff
+     * transfer_operation_flag is PLDM_ACKNOWLEDGEMENT_ONLY and
+     * event_id_to_acknowledge is not PLDM_PLATFORM_EVENT_ID_FRAGMENT
      */
-    requestMsg[4] = PLDM_GET_FIRSTPART;
-    requestMsg[9] = 0xff;
-    requestMsg[10] = 0xff;
+    requestMsg[4] = PLDM_ACKNOWLEDGEMENT_ONLY;
+    requestMsg[9] = 0x11;
+    requestMsg[10] = 0x22;
 
     rc = decode_poll_for_platform_event_message_req(
         request, requestMsg.size() - hdrSize, &retFormatVersion,
         &retTransferOperationFlag, &retDataTransferHandle,
         &retEventIdToAcknowledge);
 
-    EXPECT_EQ(rc, PLDM_ERROR_INVALID_DATA);
+    EXPECT_EQ(rc, -EPROTO);
+
+    rc = decode_poll_for_platform_event_message_req(
+        NULL, requestMsg.size() - hdrSize, &retFormatVersion,
+        &retTransferOperationFlag, &retDataTransferHandle,
+        &retEventIdToAcknowledge);
+    EXPECT_EQ(rc, -EINVAL);
+
+    rc = decode_poll_for_platform_event_message_req(
+        request, requestMsg.size() - hdrSize, nullptr,
+        &retTransferOperationFlag, &retDataTransferHandle,
+        &retEventIdToAcknowledge);
+
+    EXPECT_EQ(rc, -EINVAL);
+
+    rc = decode_poll_for_platform_event_message_req(
+        request, requestMsg.size() - hdrSize, &retFormatVersion, nullptr,
+        &retDataTransferHandle, &retEventIdToAcknowledge);
+
+    EXPECT_EQ(rc, -EINVAL);
+
+    rc = decode_poll_for_platform_event_message_req(
+        request, requestMsg.size() - hdrSize, &retFormatVersion,
+        &retTransferOperationFlag, nullptr, &retEventIdToAcknowledge);
+
+    EXPECT_EQ(rc, -EINVAL);
+
+    rc = decode_poll_for_platform_event_message_req(
+        request, requestMsg.size() - hdrSize, &retFormatVersion,
+        &retTransferOperationFlag, &retDataTransferHandle, nullptr);
+
+    EXPECT_EQ(rc, -EINVAL);
 }
 
 TEST(PollForPlatformEventMessage, testGoodEncodeResposeP1)